requirements.txt
# QorSense Backend Dependencies
# Production dependencies with pinned versions for reproducibility

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1

# Data Processing & Analysis
numpy==2.2.0
pandas==2.2.3
scipy==1.14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
redis==5.0.1
prometheus-client==0.19.0
scikit-learn==1.6.0

# Visualization & Reporting
plotly==5.24.1
kaleido==0.2.1
fpdf==1.7.2
reportlab==4.2.5

# API & HTTP
httpx==0.28.1
requests==2.32.3

# Data Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# UI (for Streamlit if needed)
streamlit==1.40.2
